enabling test for improving coverage
